libc: Move public error-handling functions to libc.c
With this patch, errno.c only contains errno-specific code. All other
code related to error handling is located in libc.c or in the directory
error/.
