+++ lib/modules/libc/src/errno.c
@@ -4,14 +4,6 @@
 
 #include "picotm/errno.h"
 #include "error/module.h"
-#include "picotm/picotm-libc.h"
-
-PICOTM_EXPORT
-void
-picotm_libc_save_errno()
-{
-    error_module_save_errno();
-}
 
 PICOTM_EXPORT
 int*
@@ -19,18 +11,4 @@ __errno_location_tx()
 {
     error_module_save_errno();
     return &errno;
-}
-
-PICOTM_EXPORT
-void
-picotm_libc_set_error_recovery(enum picotm_libc_error_recovery recovery)
-{
-    return error_module_set_error_recovery(recovery);
-}
-
-PICOTM_EXPORT
-enum picotm_libc_error_recovery
-picotm_libc_get_error_recovery()
-{
-    return error_module_get_error_recovery();
 }
