+++ lib/modules/libc/src/libc.c
@@ -3,6 +3,36 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "picotm/picotm-libc.h"
+#include "error/module.h"
+
+/*
+ * Error handling
+ */
+
+PICOTM_EXPORT
+void
+picotm_libc_save_errno()
+{
+    error_module_save_errno();
+}
+
+PICOTM_EXPORT
+void
+picotm_libc_set_error_recovery(enum picotm_libc_error_recovery recovery)
+{
+    return error_module_set_error_recovery(recovery);
+}
+
+PICOTM_EXPORT
+enum picotm_libc_error_recovery
+picotm_libc_get_error_recovery()
+{
+    return error_module_get_error_recovery();
+}
+
+/*
+ * File I/O
+ */
 
 static enum picotm_libc_cc_mode g_file_type_cc_mode[] = {
     PICOTM_LIBC_CC_MODE_NOUNDO,
